Add custom file filter
If we make GitHub Action jobs required in order to merge, then on every PR GH automatically creates a status check in the pending state. This means that this workflow needs to run in order to update the status checks. I cannot use the GH workflow path filter because the job would be skipped.